docu: put .txt file contents in a <pre>
Instead of minimally-viably styling contents of .txt files, simply leave
them as-is, and put them into a <pre> element instead.

Thinking here is that this is more in line with the nature of .txt
files, where the author has full control over spacing, layout etc
through the affordance that a monospace text grid provides. If actual
styling is desired, the other formats support that better.

To aid various combinations of small viewports and ultra-wide content,
we still apply text wrapping.